requirements.txt
fastapi
uvicorn
python-dotenv
selenium
webdriver-manager
